Refactor: 에러 발생시 throw new Error
